v2 - broke the generator into multiple modules
